41725-fork-align-gcf/pyproject.toml
[project]
name = "repro-gcf"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "functions-framework==3.10.1",
    "grpcio==1.78.0",
    # "grpcio==1.78.1",
    "grpcio_tools",
]
